create bits and genre model
