+++ BitHub/BitHub.csproj
@@ -177,9 +177,20 @@
     <Compile Include="Migrations\201905210848456_InitialModel.Designer.cs">
       <DependentUpon>201905210848456_InitialModel.cs</DependentUpon>
     </Compile>
+    <Compile Include="Migrations\201905210909570_createBit.cs" />
+    <Compile Include="Migrations\201905210909570_createBit.Designer.cs">
+      <DependentUpon>201905210909570_createBit.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Migrations\201905210930120_addConventions.cs" />
+    <Compile Include="Migrations\201905210930120_addConventions.Designer.cs">
+      <DependentUpon>201905210930120_addConventions.cs</DependentUpon>
+    </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
-    <Compile Include="Models\IdentityModels.cs" />
+    <Compile Include="Models\ApplicationDbContext.cs" />
+    <Compile Include="Models\ApplicationUser.cs" />
+    <Compile Include="Models\Bit.cs" />
+    <Compile Include="Models\Genre.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
@@ -260,6 +271,12 @@
   <ItemGroup>
     <EmbeddedResource Include="Migrations\201905210848456_InitialModel.resx">
       <DependentUpon>201905210848456_InitialModel.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201905210909570_createBit.resx">
+      <DependentUpon>201905210909570_createBit.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201905210930120_addConventions.resx">
+      <DependentUpon>201905210930120_addConventions.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <PropertyGroup>
