+++ BitHub/Models/ApplicationDbContext.cs
@@ -1,0 +1,21 @@
+﻿using System.Data.Entity;
+using Microsoft.AspNet.Identity.EntityFramework;
+
+namespace BitHub.Models
+{
+    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+    {
+        public DbSet<Bit> Bits { get; set; }
+        public DbSet<Genre> Genres { get; set; }
+
+        public ApplicationDbContext()
+            : base("DefaultConnection", throwIfV1Schema: false)
+        {
+        }
+
+        public static ApplicationDbContext Create()
+        {
+            return new ApplicationDbContext();
+        }
+    }
+}
