+++ BitHub/Models/Bit.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace BitHub.Models
+{
+    public class Bit
+    {
+        public int Id { get; set; }
+        [Required]
+        public ApplicationUser Artist { get; set; }
+        public DateTime Date { get; set; }
+        [Required]
+        [StringLength(255)]
+        public string Venue { get; set; }
+        [Required]
+        public Genre Genre { get; set; }
+    }
+}
